refactor(cli): extract event renderers
